Changed:
- Changed all code for support with REST and spliced them
- Added GSON Serialisers for sending appropriate data
- Added comments
- Deleted unneeded code
- Made A LOT of bug fixes
- Made sure that filters are activated appropriate
- Changed A LOT of HTML content
- Deleted some unneeded JQueries
- Fixed HTML JQUERY

Logout not working anymore, need to look at that..

+++ Website/ProjectVirt/src/beans/VMBean.java
@@ -14,7 +14,7 @@ public class VMBean {
 	/**
 	 * Setting the variables needed
 	 */
-	private int vmID;
+	private long vmID;
 	private String vmName;
 	private String vmCPU;
 	private String vmMemory;
@@ -132,11 +132,11 @@ public class VMBean {
 		valid = newValid;
 	}
 
-	public int getVmID() {
+	public long getVmID() {
 		return vmID;
 	}
 
-	public void setVmID(int vmID) {
+	public void setVmID(long vmID) {
 		this.vmID = vmID;
 	}
